+++ src/io/pajek-header.h
@@ -34,7 +34,7 @@
  */
 #ifdef FUZZING_BUILD_MODE_UNSAFE_FOR_PRODUCTION
 /* Limit maximum vertex count when using a fuzzer, to avoid out-of-memory failure. */
-#define IGRAPH_PAJEK_MAX_VERTEX_COUNT (1 << 20)
+#define IGRAPH_PAJEK_MAX_VERTEX_COUNT (1 << 18)
 #else
 #define IGRAPH_PAJEK_MAX_VERTEX_COUNT INT32_MAX
 #endif