+++ src/io/pajek.c
@@ -29,6 +29,8 @@
 
 #include "graph/attributes.h"
 
+#include "internal/hacks.h" /* IGRAPH_STATIC_ASSERT */
+
 #include "io/pajek-header.h"
 #include "io/parsers/pajek-parser.h"
 
@@ -77,14 +79,14 @@ void igraph_i_pajek_destroy_attr_vector(igraph_vector_ptr_t *attrs) {
  * \brief Reads a file in Pajek format.
  *
  * Only a subset of the Pajek format is implemented. This is partially
- * because this format is not fully documented, but also because
+ * because there is no formal specification for this format, but also because
  * <command>igraph</command> does not support some Pajek features, like
  * mixed graphs.
  *
  * </para><para>
  * Starting from version 0.6.1 igraph reads bipartite (two-mode)
- * graphs from Pajek files and add the \c type vertex attribute for them.
- * Warnings are given for invalid edges, i.e. edges connecting
+ * graphs from Pajek files and adds the \c type Boolean vertex attribute for
+ * them. Warnings are given for invalid edges, i.e. edges connecting
  * vertices of the same type.
  *
  * </para><para>
@@ -92,14 +94,15 @@ void igraph_i_pajek_destroy_attr_vector(igraph_vector_ptr_t *attrs) {
  * \olist
  * \oli Only <filename>.net</filename> files are supported, Pajek
  * project files (<filename>.paj</filename>) are not.
- * \oli Time events networks are not supported.
- * \oli Hypergraphs (i.e. graphs with non-binary edges) are not
- * supported.
+ * \oli Temporal networks (i.e. with time events) are not supported.
  * \oli Graphs with both directed and non-directed edges are not
  * supported, as they cannot be represented in <command>igraph</command>.
  * \oli Only Pajek networks are supported; permutations, hierarchies,
  * clusters and vectors are not.
- * \oli Graphs with multiple edge sets are not supported.
+ * \oli Multi-relational networks (i.e. networks with multiple edge
+ * types) are not supported.
+ * \oli Unicode characters encoded as <code>&amp;#dddd;</code>, or newlines
+ * encoded as <code>\n</code> will not be decoded.
  * \endolist
  *
  * </para><para>
@@ -111,7 +114,7 @@ void igraph_i_pajek_destroy_attr_vector(igraph_vector_ptr_t *attrs) {
  * \c labeldegree2 instead of \c lphi, \c framewidth instead of \c bw,
  * \c fontsize instead of \c fos, \c rotation instead of \c phi,
  * \c radius instead of \c r, \c diamondratio instead of \c q,
- * \c labeldegree instead of \c la, \c vertexsize instead of \c size,
+ * \c labeldegree instead of \c la,
  * \c color instead of \c ic, \c framecolor instead of \c bc,
  * \c labelcolor instead of \c lc; these belong to vertices.
  *
@@ -133,11 +136,13 @@ void igraph_i_pajek_destroy_attr_vector(igraph_vector_ptr_t *attrs) {
  *
  * </para><para>
  * In addition the following vertex attributes might be added: \c id
- * if there are vertex IDs in the file, \c x and \c y or \c x
- * and \c y and \c z if there are vertex coordinates in the file.
+ * and \c name are added (with the same value) if there are vertex IDs in the
+ * file. \c id is deprecated in favour of \c name and will no longer be used
+ * by future versions of igraph. \c x and \c y, and potentially \c z are also
+ * added if there are vertex coordinates in the file.
  *
- * </para><para>The \c weight edge attribute might be
- * added if there are edge weights present.
+ * </para><para>
+ * The \c weight edge attribute will be added if there are edge weights present.
  *
  * </para><para>
  * See the Pajek homepage:
@@ -145,7 +150,9 @@ void igraph_i_pajek_destroy_attr_vector(igraph_vector_ptr_t *attrs) {
  * Pajek. The Pajek manual,
  * http://vlado.fmf.uni-lj.si/pub/networks/pajek/doc/pajekman.pdf,
  * and http://mrvar.fdv.uni-lj.si/pajek/DrawEPS.htm
- * have information on the Pajek file format.
+ * have information on the Pajek file format. There is additional
+ * useful information and sample files at
+ * http://mrvar.fdv.uni-lj.si/pajek/history.htm
  *
  * \param graph Pointer to an uninitialized graph object.
  * \param file An already opened file handler.
@@ -284,56 +291,26 @@ igraph_error_t igraph_read_graph_pajek(igraph_t *graph, FILE *instream) {
 #define V_SHAPE             4
 #define V_XFACT             5
 #define V_YFACT             6
-#define V_COLOR_RED         7
-#define V_COLOR_GREEN       8
-#define V_COLOR_BLUE        9
-#define V_FRAMECOLOR_RED   10
-#define V_FRAMECOLOR_GREEN 11
-#define V_FRAMECOLOR_BLUE  12
-#define V_LABELCOLOR_RED   13
-#define V_LABELCOLOR_GREEN 14
-#define V_LABELCOLOR_BLUE  15
-#define V_LABELDIST        16
-#define V_LABELDEGREE2     17
-#define V_FRAMEWIDTH       18
-#define V_FONTSIZE         19
-#define V_ROTATION         20
-#define V_RADIUS           21
-#define V_DIAMONDRATIO     22
-#define V_LABELDEGREE      23
-#define V_VERTEXSIZE       24
-#define V_FONT             25
-#define V_URL              26
-#define V_COLOR            27
-#define V_FRAMECOLOR       28
-#define V_LABELCOLOR       29
-#define V_LAST             30
+#define V_LABELDIST         7
+#define V_LABELDEGREE2      8
+#define V_FRAMEWIDTH        9
+#define V_FONTSIZE         10
+#define V_ROTATION         11
+#define V_RADIUS           12
+#define V_DIAMONDRATIO     13
+#define V_LABELDEGREE      14
+#define V_FONT             15
+#define V_URL              16
+#define V_COLOR            17
+#define V_FRAMECOLOR       18
+#define V_LABELCOLOR       19
+#define V_LAST             20
 
 #define E_WEIGHT            0
-#define E_COLOR_RED         1
-#define E_COLOR_GREEN       2
-#define E_COLOR_BLUE        3
-#define E_ARROWSIZE         4
-#define E_EDGEWIDTH         5
-#define E_HOOK1             6
-#define E_HOOK2             7
-#define E_ANGLE1            8
-#define E_ANGLE2            9
-#define E_VELOCITY1        10
-#define E_VELOCITY2        11
-#define E_ARROWPOS         12
-#define E_LABELPOS         13
-#define E_LABELANGLE       14
-#define E_LABELANGLE2      15
-#define E_LABELDEGREE      16
-#define E_FONTSIZE         17
-#define E_ARROWTYPE        18
-#define E_LINEPATTERN      19
-#define E_LABEL            20
-#define E_LABELCOLOR       21
-#define E_COLOR            22
-#define E_LAST             23
+#define E_LAST              1
 
+/* Pajek encodes newlines as \n, and any unicode character can be encoded
+ * in the form &#hhhh;. Therefore we encode quotation marks as &#34; */
 static igraph_error_t igraph_i_pajek_escape(const char* src, char** dest) {
     igraph_integer_t destlen = 0;
     igraph_bool_t need_escape = false;
@@ -342,9 +319,12 @@ static igraph_error_t igraph_i_pajek_escape(const char* src, char** dest) {
     const char *s;
     char *d;
     for (s = src; *s; s++, destlen++) {
-        if (*s == '\\' || *s == '"' || *s == '\n' || *s == '\r') {
+        if (*s == '\n' || *s == '\r') {
             need_escape = true;
             destlen++;
+        } else if (*s == '"') {
+            need_escape = true;
+            destlen += 4;
         } else if (!isalnum(*s)) {
             need_escape = true;
         }
@@ -352,7 +332,7 @@ static igraph_error_t igraph_i_pajek_escape(const char* src, char** dest) {
 
     if (!need_escape) {
         /* At this point, we know that the string does not contain any chars
-         * that would warrant escaping. Therefore, we simply quote it and
+         * that would warrant encoding. Therefore, we simply quote it and
          * return the quoted string. This is necessary because Pajek uses some
          * reserved words in its format (like 'c' standing for color) and they
          * have to be quoted as well.
@@ -375,16 +355,18 @@ static igraph_error_t igraph_i_pajek_escape(const char* src, char** dest) {
 
     for (s = src; *s; s++, d++) {
         switch (*s) {
-        case '\\':
+        /* Encode quotation marks as &#34;, as they would otherwise signify
+           the end/beginning of a string. */
         case '"':
-            *d = '\\'; d++;
-            *d = *s;
+            strcpy(d, "&#34;"); d += 4; break;
             break;
-        /* Convert both CR and LF to "\n", as neither should apear in a quoted string. */
+        /* Encode both CR and LF as \n, as neither should apear in a quoted string.
+           \n is the _only_ escape sequence Pajek understands. */
         case '\n':
         case '\r':
             *d = '\\'; d++;
             *d = 'n';
+            break;
         default:
             *d = *s;
         }
@@ -404,11 +386,15 @@ static igraph_error_t igraph_i_pajek_escape(const char* src, char** dest) {
  *
  * </para><para>
  * The Pajek vertex and edge parameters (like color) are determined by
- * the attributes of the vertices and edges, of course this requires
+ * the attributes of the vertices and edges. Of course this requires
  * an attribute handler to be installed. The names of the
  * corresponding vertex and edge attributes are listed at \ref
  * igraph_read_graph_pajek(), e.g. the \c color vertex attributes
  * determines the color (\c c in Pajek) parameter.
+ *
+ * </para><para>
+ * Vertex and edge attributes that do not correspond to any documented
+ * Pajek parameter are discarded.
  *
  * </para><para>
  * As of version 0.6.1 igraph writes bipartite graphs into Pajek files
@@ -431,11 +417,14 @@ static igraph_error_t igraph_i_pajek_escape(const char* src, char** dest) {
  * line endings manually with a text editor or with \c unix2dos or \c iconv
  * from the command line).
  *
+ * </para><para>
+ * Pajek will only interpret UTF-8 encoded files if they contain a byte-order
+ * mark (BOM) at the beginning. igraph is agnostic of string attribute encodings
+ * and therefore it will never write a BOM. You need to add this manually
+ * if/when necessary.
+ *
  * \param graph The graph object to write.
- * \param outstream The file to write to. It should be opened and
- * writable. Make sure that you open the file in binary format if you use MS Windows,
- * otherwise end of line characters will be messed up. (igraph will be able
- * to read back these messed up files, but Pajek won't.)
+ * \param outstream The file to write to. It should be opened and writable.
  * \return Error code.
  *
  * Time complexity: O(|V|+|E|+|A|), |V| is the number of vertices, |E|
@@ -451,42 +440,42 @@ static igraph_error_t igraph_i_pajek_escape(const char* src, char** dest) {
 
 igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) {
     igraph_integer_t no_of_nodes = igraph_vcount(graph);
-    igraph_integer_t i, j;
 
     igraph_attribute_type_t vtypes[V_LAST], etypes[E_LAST];
     igraph_bool_t write_vertex_attrs = false;
 
     /* Same order as the #define's */
     const char *vnames[] = { "id", "x", "y", "z", "shape", "xfact", "yfact",
-                             "", "", "", "", "", "", "", "", "",
-                             "labeldist", "labeldegree2", "framewidth",
-                             "fontsize", "rotation", "radius",
-                             "diamondratio", "labeldegree", "vertexsize",
-                             "font", "url", "color", "framecolor",
-                             "labelcolor"
-                           };
+        "labeldist", "labeldegree2", "framewidth",
+        "fontsize", "rotation", "radius",
+        "diamondratio", "labeldegree",
+        "font", "url", "color", "framecolor",
+        "labelcolor"
+    };
+    IGRAPH_STATIC_ASSERT(sizeof(vnames) / sizeof(vnames[0]) == V_LAST);
 
+    /* Arrays called xxx[]  are igraph attribute names,
+     *               xxx2[] are the corresponding Pajek names. */
     const char *vnumnames[] = { "xfact", "yfact", "labeldist",
                                 "labeldegree2", "framewidth", "fontsize",
                                 "rotation", "radius", "diamondratio",
-                                "labeldegree", "vertexsize"
+                                "labeldegree"
                               };
     const char *vnumnames2[] = { "x_fact", "y_fact", "lr", "lphi", "bw",
-                                 "fos", "phi", "r", "q", "la", "size"
+                                 "fos", "phi", "r", "q", "la"
                                };
+    IGRAPH_STATIC_ASSERT(sizeof(vnumnames) == sizeof(vnumnames2));
+
     const char *vstrnames[] = { "font", "url", "color", "framecolor",
                                 "labelcolor"
                               };
     const char *vstrnames2[] = { "font", "url", "ic", "bc", "lc" };
+    IGRAPH_STATIC_ASSERT(sizeof(vstrnames) == sizeof(vstrnames2));
 
-    const char *enames[] = { "weight", "", "", "",
-                             "arrowsize", "edgewidth", "hook1", "hook2",
-                             "angle1", "angle2", "velocity1", "velocity2",
-                             "arrowpos", "labelpos", "labelangle",
-                             "labelangle2", "labeldegree", "fontsize",
-                             "arrowtype", "linepattern", "label", "labelcolor",
-                             "color"
-                           };
+    /* Same order as the #define's */
+    const char *enames[] = { "weight" };
+    IGRAPH_STATIC_ASSERT(sizeof(enames) / sizeof(enames[0]) == E_LAST);
+
     const char *enumnames[] = { "arrowsize", "edgewidth", "hook1", "hook2",
                                 "angle1", "angle2", "velocity1", "velocity2",
                                 "arrowpos", "labelpos", "labelangle",
@@ -495,10 +484,13 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
     const char *enumnames2[] = { "s", "w", "h1", "h2", "a1", "a2", "k1", "k2",
                                  "ap", "lp", "lr", "lphi", "la", "fos"
                                };
+    IGRAPH_STATIC_ASSERT(sizeof(enumnames) == sizeof(enumnames2));
+
     const char *estrnames[] = { "arrowtype", "linepattern", "label",
-                                "labelcolor", "color"
+                                "labelcolor", "color", "font"
                               };
-    const char *estrnames2[] = { "a", "p", "l", "lc", "c" };
+    const char *estrnames2[] = { "a", "p", "l", "lc", "c", "font" };
+    IGRAPH_STATIC_ASSERT(sizeof(estrnames) == sizeof(estrnames2));
 
     /* Newer Pajek versions support both Unix and Windows-style line endings,
      * so we just use Unix style. This will get converted to CRLF on Windows
@@ -532,22 +524,18 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
     IGRAPH_VECTOR_INT_INIT_FINALLY(&vx_numa, 0);
     IGRAPH_VECTOR_INT_INIT_FINALLY(&vx_stra, 0);
 
-    /* Check if graph is bipartite */
+    /* Check if graph is bipartite, i.e. whether it has a Boolean 'type' vertex attribute. */
     if (igraph_i_attribute_has_attr(graph, IGRAPH_ATTRIBUTE_VERTEX, "type")) {
         igraph_attribute_type_t type_type;
         IGRAPH_CHECK(igraph_i_attribute_gettype(graph, &type_type, IGRAPH_ATTRIBUTE_VERTEX, "type"));
         if (type_type == IGRAPH_ATTRIBUTE_BOOLEAN) {
-            igraph_integer_t bptr = 0, tptr = 0;
             bipartite = true; write_vertex_attrs = true;
             /* Count top and bottom vertices, we go over them twice,
             because we want to keep their original order */
-            IGRAPH_CHECK(igraph_vector_int_init(&bip_index, no_of_nodes));
-            IGRAPH_FINALLY(igraph_vector_int_destroy, &bip_index);
-            IGRAPH_CHECK(igraph_vector_int_init(&bip_index2, no_of_nodes));
-            IGRAPH_FINALLY(igraph_vector_int_destroy, &bip_index2);
-            IGRAPH_CHECK(igraph_vector_bool_init(&bvec, 1));
-            IGRAPH_FINALLY(igraph_vector_bool_destroy, &bvec);
-            for (i = 0; i < no_of_nodes; i++) {
+            IGRAPH_VECTOR_INT_INIT_FINALLY(&bip_index, no_of_nodes);
+            IGRAPH_VECTOR_INT_INIT_FINALLY(&bip_index2, no_of_nodes);
+            IGRAPH_VECTOR_BOOL_INIT_FINALLY(&bvec, 1);
+            for (igraph_integer_t i = 0; i < no_of_nodes; i++) {
                 IGRAPH_CHECK(igraph_i_attribute_get_bool_vertex_attr(graph,
                              "type", igraph_vss_1(i), &bvec));
                 if (VECTOR(bvec)[0]) {
@@ -556,7 +544,7 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
                     nobottom++;
                 }
             }
-            for (i = 0, bptr = 0, tptr = nobottom; i < no_of_nodes; i++) {
+            for (igraph_integer_t i = 0, bptr = 0, tptr = nobottom; i < no_of_nodes; i++) {
                 IGRAPH_CHECK(igraph_i_attribute_get_bool_vertex_attr(graph,
                              "type", igraph_vss_1(i), &bvec));
                 if (VECTOR(bvec)[0]) {
@@ -586,9 +574,9 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
         }
     }
 
-    /* Check the vertex attributes */
+    /* Check the vertex attributes, and determine if we need to write them. */
     memset(vtypes, 0, sizeof(vtypes[0])*V_LAST);
-    for (i = 0; i < V_LAST; i++) {
+    for (igraph_integer_t i = 0; i < V_LAST; i++) {
         if (igraph_i_attribute_has_attr(graph, IGRAPH_ATTRIBUTE_VERTEX, vnames[i])) {
             IGRAPH_CHECK(igraph_i_attribute_gettype(
                              graph, &vtypes[i], IGRAPH_ATTRIBUTE_VERTEX, vnames[i]));
@@ -597,7 +585,7 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
             vtypes[i] = (igraph_attribute_type_t) -1;
         }
     }
-    for (i = 0; i < (igraph_integer_t) (sizeof(vnumnames) / sizeof(vnumnames[0])); i++) {
+    for (igraph_integer_t i = 0; i < (igraph_integer_t) (sizeof(vnumnames) / sizeof(vnumnames[0])); i++) {
         igraph_attribute_type_t type;
         if (igraph_i_attribute_has_attr(graph, IGRAPH_ATTRIBUTE_VERTEX, vnumnames[i])) {
             IGRAPH_CHECK(igraph_i_attribute_gettype(
@@ -607,7 +595,7 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
             }
         }
     }
-    for (i = 0; i < (igraph_integer_t) (sizeof(vstrnames) / sizeof(vstrnames[0])); i++) {
+    for (igraph_integer_t i = 0; i < (igraph_integer_t) (sizeof(vstrnames) / sizeof(vstrnames[0])); i++) {
         igraph_attribute_type_t type;
         if (igraph_i_attribute_has_attr(graph, IGRAPH_ATTRIBUTE_VERTEX, vstrnames[i])) {
             IGRAPH_CHECK(igraph_i_attribute_gettype(
@@ -620,7 +608,7 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
 
     /* Write vertices */
     if (write_vertex_attrs) {
-        for (i = 0; i < no_of_nodes; i++) {
+        for (igraph_integer_t i = 0; i < no_of_nodes; i++) {
             igraph_integer_t id = bipartite ? VECTOR(bip_index)[i] : i;
 
             /* vertex ID */
@@ -672,7 +660,7 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
             }
 
             /* numeric parameters */
-            for (j = 0; j < igraph_vector_int_size(&vx_numa); j++) {
+            for (igraph_integer_t j = 0; j < igraph_vector_int_size(&vx_numa); j++) {
                 igraph_integer_t idx = VECTOR(vx_numa)[j];
                 IGRAPH_CHECK(igraph_i_attribute_get_numeric_vertex_attr(
                                  graph, vnumnames[idx], igraph_vss_1(id), &numv));
@@ -681,7 +669,7 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
             }
 
             /* string parameters */
-            for (j = 0; j < igraph_vector_int_size(&vx_stra); j++) {
+            for (igraph_integer_t j = 0; j < igraph_vector_int_size(&vx_stra); j++) {
                 igraph_integer_t idx = VECTOR(vx_stra)[j];
                 IGRAPH_CHECK(igraph_i_attribute_get_string_vertex_attr(
                                  graph, vstrnames[idx], igraph_vss_1(id), &strv));
@@ -709,7 +697,8 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
     IGRAPH_FINALLY(igraph_eit_destroy, &eit);
 
     /* Check edge attributes */
-    for (i = 0; i < E_LAST; i++) {
+    /* TODO: refactor and simplify since only "weight" is relevant */
+    for (igraph_integer_t i = 0; i < E_LAST; i++) {
         if (igraph_i_attribute_has_attr(graph, IGRAPH_ATTRIBUTE_EDGE, enames[i])) {
             IGRAPH_CHECK(igraph_i_attribute_gettype(
                              graph, &etypes[i], IGRAPH_ATTRIBUTE_EDGE, enames[i]));
@@ -717,7 +706,7 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
             etypes[i] = (igraph_attribute_type_t) -1;
         }
     }
-    for (i = 0; i < (igraph_integer_t) (sizeof(enumnames) / sizeof(enumnames[0])); i++) {
+    for (igraph_integer_t i = 0; i < (igraph_integer_t) (sizeof(enumnames) / sizeof(enumnames[0])); i++) {
         igraph_attribute_type_t type;
         if (igraph_i_attribute_has_attr(graph, IGRAPH_ATTRIBUTE_EDGE, enumnames[i])) {
             IGRAPH_CHECK(igraph_i_attribute_gettype(
@@ -727,7 +716,7 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
             }
         }
     }
-    for (i = 0; i < (igraph_integer_t) (sizeof(estrnames) / sizeof(estrnames[0])); i++) {
+    for (igraph_integer_t i = 0; i < (igraph_integer_t) (sizeof(estrnames) / sizeof(estrnames[0])); i++) {
         igraph_attribute_type_t type;
         if (igraph_i_attribute_has_attr(graph, IGRAPH_ATTRIBUTE_EDGE, estrnames[i])) {
             IGRAPH_CHECK(igraph_i_attribute_gettype(
@@ -738,7 +727,7 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
         }
     }
 
-    for (i = 0; !IGRAPH_EIT_END(eit); IGRAPH_EIT_NEXT(eit), i++) {
+    for (igraph_integer_t i = 0; !IGRAPH_EIT_END(eit); IGRAPH_EIT_NEXT(eit), i++) {
         igraph_integer_t edge = IGRAPH_EIT_GET(eit);
         igraph_integer_t from, to;
         igraph_edge(graph, edge, &from,  &to);
@@ -757,7 +746,7 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
         }
 
         /* numeric parameters */
-        for (j = 0; j < igraph_vector_int_size(&ex_numa); j++) {
+        for (igraph_integer_t j = 0; j < igraph_vector_int_size(&ex_numa); j++) {
             igraph_integer_t idx = VECTOR(ex_numa)[j];
             IGRAPH_CHECK(igraph_i_attribute_get_numeric_edge_attr(
                              graph, enumnames[idx], igraph_ess_1(edge), &numv));
@@ -766,7 +755,7 @@ igraph_error_t igraph_write_graph_pajek(const igraph_t *graph, FILE *outstream) 
         }
 
         /* string parameters */
-        for (j = 0; j < igraph_vector_int_size(&ex_stra); j++) {
+        for (igraph_integer_t j = 0; j < igraph_vector_int_size(&ex_stra); j++) {
             igraph_integer_t idx = VECTOR(ex_stra)[j];
             IGRAPH_CHECK(igraph_i_attribute_get_string_edge_attr(
                              graph, estrnames[idx], igraph_ess_1(edge), &strv));