+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## [master]
 
+### Added
+
+ - `igraph_joint_degree_matrix()` computes the joint degree matrix, i.e. counts connections between vertices of different degrees. (PR #2407 by Lára Margrét Hólmfríðardóttir @larah19)
+
 ### Changes
 
- - The behaviour of the Pajek format reader is now more closely aligned with the Pajek software and the reader is more tolerant of input it cannot interpret.
+ - The behaviour of the Pajek format reader and writer is now more closely aligned with the Pajek software and the reader is more tolerant of input it cannot interpret. Only those vertex and edge parameters are treated as valid which Pajek itself understands, therefore support for `size` is now dropped, and support for the `font` edge parameter is added. See http://mrvar.fdv.uni-lj.si/pajek/DrawEPS.htm for more information. Invalid/unrecognized parameters are now converted to igraph attributes by the reader, but just as before, they are not output by the writer.
+ - The Pajek format writer now encodes newline and quotation mark characters in a Pajek-compatible manner (`\n` and `&#34;`, respectively).
 
 ### Fixed
 
@@ -15,6 +20,7 @@
 ### Other
 
  - Updated vendored mini-gmp to 6.3.0.
+ - `igraph_connected_components()` makes better use of the cache, improving overall performance.
  - Documentation improvements.
 
 ## [0.10.7] - 2023-09-04